@@ -2,13 +2,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <Company>Internet Sukces Piotr Stęclik</Company>
-    <Copyright>Copyright © Internet Sukces Piotr Stęclik 2016-2023</Copyright>
+    <Copyright>Copyright © Internet Sukces Piotr Stęclik 2016-2024</Copyright>
     <Product>iSukces.Code</Product>
-    <TargetFrameworks>net48;netcoreapp3.1;net6.0;net7.0;net8.0</TargetFrameworks>
+    <TargetFrameworks>net48;netcoreapp3.1;net6.0;net8.0</TargetFrameworks>
     <AssemblyVersion>1.23.1218.202</AssemblyVersion>
     <FileVersion>1.23.1218.202</FileVersion>
     <Version>1.23.1218.202</Version>
-    <LangVersion>11</LangVersion>
+    <LangVersion>12</LangVersion>
     <UseWPF>false</UseWPF>
     <SignAssembly>true</SignAssembly>
     <AssemblyOriginatorKeyFile>..\key.snk</AssemblyOriginatorKeyFile>
