@@ -1,3 +1,4 @@
+using iSukces.Code;
 using iSukces.Code.Interfaces;
 using Xunit;
 
@@ -96,21 +97,21 @@ namespace iSukces.Code.Tests.DependsOnPropertyGenerator
             {
                 switch (propertyName)
                 {
-                    case nameof(Master): return MasterDependent; // Slave,DeepSlave,EvenDeeperSlave
-                    case nameof(Slave): return ""DeepSlave,EvenDeeperSlave"";
+                    case nameof(Master): return MasterDependent; // Slave, DeepSlave, EvenDeeperSlave
+                    case nameof(Slave): return ""DeepSlave, EvenDeeperSlave"";
                     case nameof(DeepSlave): return DeepSlaveDependent; // EvenDeeperSlave
                     case nameof(FirstName):
-                    case nameof(LastName): return LastNameDependent; // FullNameA,FullNameB,FullNameA1,FullNameB1
+                    case nameof(LastName): return LastNameDependent; // FullNameA, FullNameB, FullNameA1, FullNameB1
                     case nameof(FullNameA): return ""FullNameA1"";
                 }
                 return null;
             }
 
-            public const string MasterDependent = ""Slave,DeepSlave,EvenDeeperSlave"";
+            public const string MasterDependent = ""Slave, DeepSlave, EvenDeeperSlave"";
 
             public const string DeepSlaveDependent = ""EvenDeeperSlave"";
 
-            public const string FirstNameDependent = ""FullNameA,FullNameB,FullNameA1,FullNameB1"";
+            public const string FirstNameDependent = ""FullNameA, FullNameB, FullNameA1, FullNameB1"";
 
             public const string LastNameDependent = FirstNameDependent;
 
